Move style related test to Style folder
